Using lambda for US realms
Coming to EU soon™
Downloading auction data in the US will now be much faster. And will no longer time out! I am sorry about it taking this long for me to fix it. I was made aware of it right before BFA came out. 

Decompressing gzip on client as API gateway did not want to do this for me and was a pain.