DB performance fix
Basically just reducing the number of concurrent auctions even more.
We are now down to 40. Not sure if this "performance issue" is related to Covic19, and increased load on the DB server that I'm using or now. Or if it just does not like that blizzard updates all of EU and US realms at xx:07 every hour.

I am now using a long running DB connection here as well. As that is better, just had to fix the previous issue regarding a "resolve" not being triggered for some reason in a promise. So that solves the concurrent connections issue.