Added conditional enabelign and disabeling of toField and toValue + profession dropdowns etc
Added lastModified o the dashboards and changed the width of the columns in the rule rows


Increased the max width of the dialog for the configurable dashboards


Moved the item details further in the foreground


Added default ID's for the default boards. So that it is easier to push updates to them, or notify the user of changes after they've modified theirs


Dashboards: Storage, disable checkbox, test fixes, id fix

- Stored in the indexedDB
Test fix


Test fix


Changed PERCENT default columns to NUMBER. at least for now


Set timeout for the re-calculations while editing a dashboard


Not calculating disabled boards + displaying a message when showing all boards


Started on the gold input field


Percent are now entered as 115 for 150% instead of 1,15 etc.

New dashboards are now added to the beginning of the list as well.
Deleting dashboards


iterable rules support basic math in their rules also for a field


Added shadowlands to the expansion  list


Various fixes, quality locales, rule component toValue handles itemSubClass and quality etc


Tooltip fix


Starting with the migration script


Work on the dashboard migration


Changed the news component to use the dialog component etc++


Migrated item details to use the dialog component.

Much better!
.


reusable dialog button


More migration


Last changes