Added the rest of the shadowlands reputations + bug fix
Reputations:
Added the rest of the reputations for shadowlands based on the wowhead tutorial for reputations. If I find any to be missing, Il'l update it in the future.

Initial load:
Fix in case of missing values for dashboard upon restore.

Hotfix:
Found an issue when restoring saved dashboards that did not have rules defined.