Added support for classic TBC auction data
Classic TBC:
Since Blizzard has released their auction API for classic TBC, I've finally added support for classic.

The realms in the dropdowns now has a suffix for "retail" or "classic". You can change to either of these realms the same way as before in the top right corner or in the settings.

For now, if you want to change to the neutral AH, go to the settings page.


- For now, recipes have not been updated  to match how they were in TBC if there has been any changes. And first aid is missing (for now).
- You can not add characters on classic, meaning that anything that filters for "known recipes" will not work. Il'l look into a solution for this as soon as I have time.

There has also ben squashed some bugs that I found in this release as well.