Removed unused indexes
You REALLY do not need to index anything that you don't need to.

Anything in the field list for each table in indexedDB, are indexed. And makes everything take a tiny bit longer to update or load.

When we load several thousand entries, this is a waste of time. Especially when you don't use the indexes for anything.